Reviewed merge sort
